requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pyhd8ed1ab_0
altair=4.2.2=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.6.2=pypi_0
bottleneck=1.3.5=py310h9128911_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.12.7=h5b45459_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=win_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py310h21b164f_0
cuda-cccl=12.1.109=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.1.105=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
dataclasses-json=0.5.7=pyhd8ed1ab_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
eigen=3.4.0=h2d74725_0
entrypoints=0.4=pypi_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.2.3=ha925a31_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.9.0=py310haa95532_0
flatbuffers=23.3.3=pypi_0
fonttools=4.39.3=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py310h2bbff1b_0
future=0.18.3=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.69.1=h5dc1a3c_2
gptcache=0.1.21=pyhd8ed1ab_0
greenlet=2.0.1=py310hd77b12b_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=nompi_he0bbb20_101
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.0=pyh025b116_0
ipython=8.12.1=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=pypi_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310h5588dad_0
kiwisolver=1.4.4=pypi_0
langchain=0.0.150=pyhd8ed1ab_0
lerc=3.0=hd77b12b_0
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.2.106=0
libcurand-dev=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.17=h8ffe710_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
llvmlite=0.39.1=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=py310h2bbff1b_0
marshmallow=3.19.0=pyhd8ed1ab_0
marshmallow-enum=1.5.1=pyh9f0ad1d_3
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
more-itertools=9.1.0=pypi_0
mpmath=1.2.1=py310haa95532_0
multidict=6.0.4=pypi_0
mypy_extensions=1.0.0=pyha770c72_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=2.8.4=py310haa95532_1
numba=0.56.4=pypi_0
numexpr=2.8.4=py310hd213c9f_0
numpy=1.23.5=pypi_0
numpy-base=1.24.3=py310h3caf3d7_0
openai=0.27.0=pypi_0
openai-whisper=20230314=pypi_0
openapi-schema-pydantic=1.2.4=pyhd8ed1ab_0
opencv=4.6.0=py310h4ed8f06_3
opencv-python=4.7.0.72=pypi_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.1=pypi_0
pandas-stubs=2.0.0.230412=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310hd77b12b_0
pip=23.0.1=py310haa95532_0
platformdirs=2.5.2=py310haa95532_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=py310h2bbff1b_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pympler=1.0.1=pypi_0
pyopenssl=23.0.0=py310haa95532_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.11=h966fe2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=2.0.0=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_3
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywin32=305=py310h2bbff1b_0
pyyaml=6.0=py310he2412df_4
pyzmq=25.0.2=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
regex=2023.3.23=pypi_0
requests=2.29.0=py310haa95532_0
rich=13.3.5=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
sentry-sdk=1.21.1=pypi_0
setuptools=66.0.0=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soundfile=0.12.1=pypi_0
sqlalchemy=1.4.39=py310h2bbff1b_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.22.0=pypi_0
stringcase=1.2.0=py_0
supervision=0.6.0=pypi_0
sympy=1.11.1=py310haa95532_0
tenacity=8.2.2=pyhd8ed1ab_0
thop=0.1.1-2209072238=pypi_0
tiktoken=0.3.1=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.0=pypi_0
torchvision=0.15.0=pypi_0
tornado=6.2=py310he2412df_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
types-pytz=2023.3.0.0=pyhd8ed1ab_0
typing-extensions=4.5.0=py310haa95532_0
typing_extensions=4.5.0=py310haa95532_0
typing_inspect=0.8.0=pyhd8ed1ab_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
ultralytics=8.0.88=pypi_0
urllib3=1.26.15=py310haa95532_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=h8ffe710_2
yarl=1.9.2=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
